Initial syntax highlighting for VS.
The VS solution at kailua_vs is tightly integrated to the Cargo project,
and with the proper Rust installation one can directly build the entire
project from VS (a bit of msbuild hack). Both shares the same `target/`
directory in addition to VS-specific `Output/` directory.

The current lexer is a bit inefficient but working. Would need further
tweak to get it to the optimal state.