Implemented error list integration and fixed a nasty memory bug.
On the memory bug, `SafeHandle` actually:

- is synchronized (surprise!);
- gets constructed by the default constructor even via P/Invoke
(thus ownsHandle should be normally true and
the handle should not be constructible otherwise);
- automatically skips ReleaseHandle call when the handle is
already invalid.

What did happen then? OwnsHandle being false means that ReleaseHandle
is never called, and the segfault happened on the Rust side. It turns
out that widestring's WideCString::into_raw is very faulty, so this
commit works around that, and now it is fully safe (thank god).
The upstream fix is on the way.