Initial implementation of roughly-working field autocompletion.
This currently only works when the checker is running in lockstep
(mostly because the subexpression detection is happening
in the checker and the error-recovering parser is still immature),
but it does work for many cases when you type slowly enough.

In the future the completion source would wait for the most recent
checker result as needed, and a common case (e.g. `NAME . NAME . |`)
will be handled much more swiftly.