Desugared function call won't generate a span for arguments anymore.
That is, `f"oo"` will only get two spans for `f` and `f"oo"`
as a whole, and `f{a=b}` will get four spans for `f`, `a`, `b` and
`f{a=b}` as a whole; `"oo"` and `{a=b}` won't get a span,
so the autocompletion can easily rely on the shortest span ending at
given cursor to display fields after `.` or `:`.

Note that keys in the table constructor will still get spans; this is
deemed to be a similar case to the confusion between l-values and
top-level non-prefix expressions.

In the resulting reorganization scheme, a new enum `Args` will
represent any possible function arguments, desugared or not.
A complete update for this change is available.