Fixed the nesting behavior of meta comments.
Previously the following example would close the nesting of
parentheses for the argument list at the newline after `--:`.

    function x(y, --: string
               z)
    end

This turned out to be a side effect of meta comments
(and more generally most "statement-level" tokens) popping
any non-statement level nesting before any action.
`--:` and (since it is hard to distinguish) a newline is
now exempted from this behavior.