Marks have been removed.
Since the VarOrCurrently flex is gone, and the VarOrConst flex is
essentially a coercion from Just which has very low probability to
turn out to be Var, we go bold and remove VarOrConst and thus all
Mark resolution codes. Technically this will require more annotation,
which can be dealt with.