Checker can now create a new field by indexing (w/ spec).
In the past with Currently, we were able to create a dummy slot for
the new field and accept the initial value into that slot in order to
handle new fields. Now we cannot do that (accept won't change
the type itself, only the flexibility changes), so the new field was
assigned but not tightly typed and causing too much problems
especially with functions and tables (both requiring tight bounds).

In the new scheme, the checker returns whether a new field was created
and a dummy slot has been assigned along with the slot (dummy if newly
created). When a new field is created a new method `Env::assign_new`
handles the situation; the prime difference is that, while acceptance
requires subtyping, declaration requires strict equality (which works
because a content of the dummy slot is always a type variable). Since
this notion clearly parallels with variable declarations, indexing can
now have the type specification, as long as the field is newly created
(otherwise it's still an error).

The old method `Checker::check_index` became internal, and now methods
`Checker::check_{l,r}val_index` should be used (along with
`Checker::assign_to_lval_index` for l-values to complete assignments).