`make_abstract` has been generalized into the general coercion.
The "coercion" now always happens when the new variable or field
gets assigned (in the slot level), or upper or exact bounds for
type variables are added (in the type level). These are the only
cases that the implicit coercion is possible.

By this commit, all known tests along with previously ignored ones
are now enabled and passes.