Revised `Env::assert_rvar_rel`.
Okay, I have to admit that I was working with faulty memory when
writing that method. I was mistaken that the row variable relationship
needs a (tail) recursion because I was trying to match them piece by
piece, but we actually had a very simple invariant ("updatable" row
variable) to cut off the recursion. This also completely got rid of
existing bugs (mainly due to the disabled tests :S).

Caveat: This does strongly rely on the fact that row variables form
a tree (children refers to the parent as a "next" row variable) and
the parent's set of positive and negative fields is a superset of
those of children. A new code does not try to heavily validate this,
so any bug can be an instant soundness hole (ugh). I believe
the current code is correct, but I'm still not 100% sure.