Non-existent fields no longer count as nil in row relations.
This was a side effect of naive row polymorphism where you can extend
rows by simply refering to the then-non-existent field (many such
cases have been now disallowed), and made `{} <: {a: string}` true.
This is no longer possible and non-existent fields can be extended
only with explicitly nilable types (currently `T?` with no type
variables, rationales follow), when the record is extended in both
sides of the row equality or in the left-hand side of the row
subtyping. Practically speaking this means that you *can* supply
additional fields to a function accepting a known record type, but
you cannot omit non-`T?` fields in them.

This change greatly affects the function calls, disjunctive implicit
unions and assignments. In most cases the left-hand side of subtyping
is implicit, while the right-hand side is explicit, so new fields are
typically added to the left-hand side---so it is much less common
that types of those fields are bare type variables (resolving to `T?`
obviously). Therefore we have deemed that the request to allow
optional fields should be explicit, and no type variables should be
resolved.