Arguments and returns cannot be omitted for non-`T?` types.
This gives much better diagnostics when arguments or returns have been
accidentally omitted. Previously we have relied on the sequence typing
but this turned out to be insufficient; so we now special case
a missing tail type (i.e. None) as a signal to more strictly check
the arity mismatch. This obviously updates lots of tests too.

Note: It is currently known that when `no_implicit_func_sig` feature
is turned off several tests are panicking. This seems to be related to
other fundamental bugs...