Implemented `--# assume C.f: method(...) --> ...` special form.
This is functionally equivalent to `--# assume static C.f:
function(self: Self, ...) --> ...` with an appropriate `Self` type.
This syntax draws a parallel to the existing method declaration syntax
in that the `self` type is implicit (in fact the code for inferring
the type for `self` has been refactored out).

This commit also makes the checker *not* to set a non-existing `x` in
`--# assume x.y` to `WHATEVER`; this is a side effect, but seems not
too wrong.