Killed secondary lookaheads.
It was thought that secondary lookaheads are used in the restricted
context and in that context they are fine, but it was just that they
were not well tested (ugh). There are two uses of secondary lookaheads
which have been replaced with condition translation from a name to
the kind, much like `Parser::convert_and_register_var_from_exp`.