Implemented delayed type checking.
Kailua normally processes the AST in the lexical order (exception is
made for `require`s however), and this has precluded checking of
mutually recursive functions. Technically it should be possible to
model them as nilable functions "assigned" at the first usage, but
this is very error-prone when used in other contexts.

This (huge) commit introduces a new pseudo-flexibility `module` to
allow for the delayed type checking of field functions or methods.
The idea is that, when the record is registered as a "module", any
function or method declarations to that module are *not* checked at
that position. Instead, they are lazily checked at the end of module-
registered scopes. Because the type checking in that scope should
proceed anyway, declarations should have their signatures explicitly
annotated (much like `[NO_CHECK]`, in fact it is internally a variant
to the `[NO_CHECK]` attribute). A new `module` flex is externally
displayed as "<initializing>", and technically is a supertype of `var`
(which makes possible to initialize `module` with other flexes).

Since this is very orthogonal to the type system, it has lots of
corner cases needed to be tested; that's why we have a separate test
file for modules alone. Other changes include:

- The indexing now correctly determines class prototypes or instances
  behind type variables.

- Flex-only specification is now allowed when appropriate.
  For example, `local x = 3 --: const` is same to
  `local x = 3 --: const integer`. This allows for the use of
  a shorter `--: module` specification instead of `--: module {...}`
  (note that `--: module {}` won't work).

- `F::Any` was unused for a long time and got removed. There is now
  a new `F::Unknown` to correctly handle the initialization of flexes
  to the field slots; they were actually incorrect but not visible
  as flex-only specification was allowed in the AST but
  not in the syntax until now.

- Lots of other refactorings necessary for the implementation.