fuzz: Fixed three bugs from the fuzzer output.
- Fixed a checker panic when it got multiple l-values and no matching
  r-values in the assignment (possible after parsing error recovery).

- Fixed a parser panic when unexpected meta block appears. (The fix is
  temporary, and we should really be able to handle them at
  any position---even though we will ultimately discard them.)

- Fixed a parser panic when `0x` is not followed with a hex digit.

This marks the beginning of the extensive fuzz testing (which was
blocked by ASAN misbehavior with libfuzzer, and I didn't realize that
I can replace ASAN with some other sanitizers until very recently).

Regression tests are now located in separate test files, as they tend
to be generated out of the context.