fuzz: Slot locks are now more fine-grained.
*sigh* So this commit tries to eliminate the very source of deadlocks
by moving the flex out of the lock and making it an atomic value. This
should make deadlocks impossible, except for one case of
`Slot::filter_by_flags` (which is flawed and should be rewritten).
Now the only way that the recursive types can fail is via real mutual
recursion.