Added support for a legacy implicit `arg` argument.
This is mainly for aiding the transition and unexpected bugs, but
the checker has been also updated to support a special vector-like
type for `arg` (formally denoted as `vector<T> & {n: integer}`, but
this is not exactly an intersection type).

This commit also gives decent errors for duplicate names in the same
scope, which Lua happily accepts (WAT). To the end, the exact
resolution depends, and currently there is at least one case that
Lua and Kailua disagrees about the name resolution in such cases.
This is known but considered acceptable for now.