Added name analysis to the parser.
Now the parser emits which tokens are names (global or local, and
corresponding scoped IDs for the latter) and what's the nature of
local names (e.g. is it an implicit `self`?). This is crucial for
IDE support like jump to definitions and context-aware renaming.

This commit also subtly changes the preference for the duplicate
function arguments: previously the specification was preferred (it
was easier to implement in that way), but it seems that the actual
Lua code is more important for IDE support (that's what the program
will eventually use anyway). For the token resolution, only matching
names will be mapped to scoped IDs on the error case---so that we
won't accidentaly rename non-matching argument names in the same
position, for example.