Moved the configuration file logic into its own crate.
A new `kailua_workspace` crate will be used both for the language
server and the CLI command to give more options to the checking.

`kailua_check::Options` interface has been changed to explicitly give
`Report` references. It was technically possible to bundle
the reporter to `Options` implementations, but it was a bit annoying,
and more importantly it didn't fit with `kailua_langsvr` diagnostics
architecture.

The format of configuration file has been updated:

- Either `kailua.json` or `.vscode/kailua.json` can be used.
  The VS Code extension will continue to use the latter for canonical,
  but the former is preferred if exists.

- `package.path` and `package.cpath` can be explicitly set at
  the configuration. In that case, setting them in the program will
  warn and have no effect (to avoid unintentional cases).

- C/C++-like comments and stray commas are now allowed (what a hack).

Finally, now having comments, the default configuration generated by
the extension is localized.