Reorganized `kailua_syntax` to reduce the names at the crate root.
Also, `kailua_syntax::{Str, Name}` are (finally) reused for `Tok`.
It had been quite annoying to manually convert `Tok::Name` to `Name`
and `Tok::Str` to `Str`, but the reorganization promotes `Name` and
`Str` types into more fundamental level and we have no reason to keep
it as is now.