Extracted `kailua_types` from `kailua_check`.
This separates the type system and operations from the checking
environment. They are pretty big components by their own, and yet
only the latter is thoroughly tested because the former required
more specialized testing mechanism and so far we couldn't easily
fit that with a larger crate. This separation will allow them to
grow and become testable individually.