Fixed a serious issue with reassignment of `module` types.
`Module` is certainly one of the most problematic feature of Kailua.
It touches every portion of other type subsystems very subtly, and
this issue is such an example.

In the current incarnation of `module` implementation it is primarily
a part of type (`F::Module` flexibility) and currently the only flex
that does not map to mutability---Kailua had eventually abandoned much
of the original flexibility design but not this. It was intended that
the type can naturally flow without special constructs, and we have
attributes at the type position for the same reason, but we had to
synchronize a list of "pending modules" (`module` slots to be handled
at the end of the current scope) with the actual `module` types.

Previously the synchronization only happened on the very mention of
`module` keywords. This proved to be erratic as `local x = y` with `y`
having `module T` will receive the `module` rightfully so but `x`
itself wouldn't be synchronized. This commit makes the synchronization
happen every time the assignment occurs; this is kinda annoying, and
we may have instead stripped `module` on assignments instead, but this
wouldn't work well with the upcoming class system revamp (where
the `[make_class]` functions will directly return `module` flex).