Class system.
This is a proper reimplementation of the half-baked and undocumented
class system found in Kailua 1.0.x. This commit completely separates
the class system logics into `ClassSystem` traits and provides
a point of extensibillity via the `--# class system` directive.
Many relics of the prior class system (e.g. `[constructor]` attribute)
have been removed, so keep in mind that this is a big breaking change.

Two class systems are currently provided. One is a "naive", unnamed
default class system that does not allow inheritance but is simpler.
Another is the `gideros` class system which strictly mimics behaviors
of the Gideros class system [1] and supports single inheritance;
it is provided as a starting point for adding various class systems.
It remains as an open question how to generalize the common traits of
many Lua class systems.

Fixes #14.

[1] http://docs.giderosmobile.com/classes_in_gideros.html